requirements.txt
amqp==5.3.1
arrow==1.4.0
asgiref==3.10.0
billiard==4.2.3
binaryornot==0.4.4
celery==5.5.3
certifi==2025.11.12
chardet==5.2.0
charset-normalizer==3.4.4
click==8.3.1
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
colorama==0.4.6
cookiecutter==2.6.0
Django==5.2.8
django-cors-headers==4.9.0
django-filter==25.2
django-tailwind==4.4.1
djangorestframework==3.16.1
et_xmlfile==2.0.0
idna==3.11
Jinja2==3.1.6
kombu==5.5.4
markdown-it-py==4.0.0
MarkupSafe==3.0.3
mdurl==0.1.2
numpy==2.3.5
openpyxl==3.1.5
packaging==25.0
pandas==2.3.3
pillow==12.0.0
prompt_toolkit==3.0.52
Pygments==2.19.2
pytailwindcss==0.3.0
python-dateutil==2.9.0.post0
python-decouple==3.8
python-slugify==8.0.4
pytz==2025.2
PyYAML==6.0.3
redis==7.0.1
reportlab==4.4.5
requests==2.32.5
rich==14.2.0
six==1.17.0
sqlparse==0.5.3
text-unidecode==1.3
tzdata==2025.2
urllib3==2.5.0
vine==5.1.0
wcwidth==0.2.14
